Preserve fractional loss delays until timer scheduling
Keep the time-threshold loss delay fractional, with a 1ms minimum, and round only the final deadline. Update regression coverage to catch an extra tick caused by rounding the duration first.

Remove unnecessary integer conversions from the ECN check and explicit casts around ceil results in the library and tests.

+++ t/test.c
@@ -1306,7 +1306,7 @@ static void test_fractional_close_timeout(void)
         quicly_conn_t *conn = draining ? server : client;
         quicly_stats_t stats;
         quicly_get_stats(conn, &stats);
-        int64_t expires_at = (int64_t)ceil(
+        int64_t expires_at = ceil(
             quic_now + quic_now_submillisec +
             4 * quicly_rtt_get_pto(&stats.rtt, quicly_get_remote_transport_parameters(conn)->max_ack_delay, quic_ctx.loss.min_pto));
         ok(quicly_get_first_timeout(conn) == expires_at);
@@ -1350,7 +1350,7 @@ static void test_fractional_rtt_measurement(void)
     quicly_get_stats(client, &stats);
     double pto =
         quicly_rtt_get_pto(&stats.rtt, quicly_get_remote_transport_parameters(client)->max_ack_delay, quic_ctx.loss.min_pto);
-    ok(quicly_get_first_timeout(client) == (int64_t)ceil(quic_now + quic_now_submillisec + pto));
+    ok(quicly_get_first_timeout(client) == ceil(quic_now + quic_now_submillisec + pto));
     ok(decode_packets(&decoded, &datagram, 1) == 1);
     quic_now_submillisec += 0.625;
     ok(quicly_receive(server, NULL, &fake_address.sa, &decoded) == 0);
@@ -1378,7 +1378,7 @@ static void test_fractional_rtt_measurement(void)
     num_datagrams = 1;
     ok(quicly_send(client, &dest, &src, &datagram, &num_datagrams, buf, sizeof(buf)) == 0);
     ok(num_datagrams == 1);
-    ok(quicly_get_first_timeout(client) == (int64_t)ceil(quic_now + quic_now_submillisec + 4 * pto));
+    ok(quicly_get_first_timeout(client) == ceil(quic_now + quic_now_submillisec + 4 * pto));
 
     quicly_free(client);
     quicly_free(server);
